+++ packages/stylelint/src/index.ts
@@ -1,11 +1,26 @@
-import config from './standard'
+import base from './base'
+import { BASE_LANGUAGE_OPTIONS, MINIPROGRAM_LANGUAGE_OPTIONS } from './language-options'
+import miniprogram from './miniprogram'
+import pxtorem from './pxtorem'
+import scss from './scss'
+import uniapp from './uniapp'
+import { mergeLanguageOptions } from './utils/merge-language-options'
+import vue from './vue'
+import vueScss from './vue-scss'
+import wechatSvg from './wechat-svg'
 
-export default config
-export { config, config as standardConfig }
-export { MINIPROGRAM_LANGUAGE_OPTIONS, STANDARD_LANGUAGE_OPTIONS } from './language-options'
-export { default as minPixelConfig } from './min-pixel'
-export { default as miniprogramConfig } from './miniprogram'
-export { default as scssConfig } from './scss'
-export { mergeLanguageOptions } from './utils/merge-language-options'
-export { default as vueConfig } from './vue'
-export { default as vueScssConfig } from './vue-scss'
+export default base
+
+export {
+  base,
+  BASE_LANGUAGE_OPTIONS,
+  mergeLanguageOptions,
+  miniprogram,
+  MINIPROGRAM_LANGUAGE_OPTIONS,
+  pxtorem,
+  scss,
+  uniapp,
+  vue,
+  vueScss,
+  wechatSvg,
+}