+++ packages/stylelint/tests/fixtures/min-pixel/stylelint.config.js
@@ -1,3 +1,0 @@
-export default {
-  extends: ['@tofrankie/stylelint/standard', '@tofrankie/stylelint/min-pixel'],
-}