+++ packages/stylelint/tests/fixtures/pxtorem/stylelint.config.js
@@ -1,0 +1,3 @@
+export default {
+  extends: ['../../../dist/base.mjs', '../../../dist/pxtorem.mjs'],
+}